migliorato output

+++ src/main/java/com/planck/math/DefaultMathParser.java
@@ -33,7 +33,7 @@ public class DefaultMathParser {
         return util.eval("D(" + function + "," + variable + ")");
     }
 
-    public double calculateNumericalIntegral(IExpr function,double higherInterval, double lowerInterval, String forVariable) {
+    public double calculateNumericalIntegral(IExpr function,int higherInterval, int lowerInterval, String forVariable) {
         util.clearVariables();
         return util.eval("NIntegrate(" + function.toString() + ", {"+ forVariable +", "+ lowerInterval +", "+ higherInterval+"})").evalDouble();
     }
